requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
requests
pdf2docx
pandas
numpy
openpyxl
streamlit
